requirements.txt
ccxt>=2.0.0
PyYAML
loguru
flask
cryptofeed
psutil
